+++ CHANGELOG.md
@@ -5,6 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0]
+
+### Added
+- chrome web store link in "About" page
+- Course File button
+
+### Changed
+- In Popup page, year to year input changed to year span
+- Year span input support for both orders (left to right and right to left)
+- Popup page centered layout
+
+### Fixed
+- "About" page not fully displaying on new user setup
+- Converting course names working properly when a course number is invalid.
+
 ## [1.3.0]
 
 ### Added